requirements.txt
playwright
beautifulsoup4
lxml
jinja2
pandas
sendgrid
certifi
python-dotenv